Add output functionality to FT Tool
This commit introduces a new abstract method for tasks. Tasks will have
an "output()" function which saves the results of the run in a machine
readable format such that it may be consumed by other modules (i.e.
scoring component to determine the pass/failure rate of the tests,
dashboard to display results).

Additionally output() should handle printing any human-read friendly
output to the stdout logs.

Signed-off-by: Salvatore Daniele <sdaniele@redhat.com>