Add Flow Test result evaluator
This commit introduces a new module evaluator which takes a path to tft
logs and evaluates performance. The threshold for pass/failure can be
set via a config file (example provided). The class will generate a
machine-readable file containing the results of the flow tests
(pass/fail).

The main driver will now be resilant again errors that occur during
runtime and return an error code if the flow tests are not completely
green.

Signed-off-by: Salvatore Daniele <sdaniele@redhat.com>