typing: fix various type annotations and specify dict/list/tuple types
Currently we run

  $ mypy --install-types --non-interactive --disallow-untyped-defs --ignore-missing-imports .

Later we will tighten this up to

  $ mypy --strict --config-file mypy.ini .

That will make these changes necessary.

There are probably some bugs here. However, in the first step it is
important to just get the linter to pass, instead of being overwhelmed
by countless warnings.  Once we have a passing linter run, we can work
on improving/fixing it further.