typing: ignore missing imports for "jc" library
This suppresses an error from mypy.

This error is a bit elusive. In certain cases, there is no error and
mypy seems to find the typing information. Sometimes however, it does
not and fails the check.

That's why I cannot just do

    import jc  # type: ignore

because there is an error if the "ignore" annotation is unneeded. And
the annotation is only needed in some unclear cases.

Anyway. This is probably a fine solution. Configure mypy to ignore
missing type definitions for "jc" module.