typing: use "|" instead of typing.Union
"|" is the same as typing.Union. Use it consistently, it's less to write
and does not require an import.

Granted, typing.Union[] has the benefit of being better searchable (with
grep). However, that is probably not very relevant. Are you ever going
to grep for functions that take/return a union?