tftbase: add tftbase module and move code from "common.py" there
There should be a clear hierarchy of modules. That is, which modules
depend on which. For example,

  - "logger.py" can only use stdlib and third party modules.
  - "host.py" can only use the above.
  - "common.py" can only use the above.

Especially the modules "logger.py", "host.py" and "common.py" contain
general purpose helpers (aside the fact that they use third party
modules). As such, they may be similarly useful in other projects. For
example, "cluster-deployment-automation" has similar modules, although
with a different implementation.

A useful (by maybe not always optimal way) of code reuse is
copy-and-paste. It should be possibly to just copy-and-paste these 3
base modules to other projects and use there with no changes. That
requires that they don't contain code specific to the project.

Move code that is only relevant for this project to a new module
"tftbase.py", where "tft" stands for "TrafficFlowTests".