all: cleanup imports
Let's not do multi line imports. It's a style choice that each import
should be on a separate line. The benefit is that the lines align (if
they are sorted) and it's easier to read. Also, a diff that adds or
removes an import, will only add/remove the line that affects the
import.

Also, group all imports to first do imports of external modules and
stdlib, then do "from" imports fro external modules and stdlib, then
imports of internal modules and finally "from" imports of internal
modules.

There is also a flake8-import-order module, but that has additional
grouping requirements. That seems too picky and the patch does not
follow its suggestions.