host: move common.Result to host.Result
This really belongs to "host" module.

Also, we should have a defined order in which modules include each other.
Currently, "host.py" imports "common.py". But that is not a good order,
because host.Host abstracts a way to call programs (on potentially other
hosts, even if that is currently not implemented). We may want to
implement helper methods based on the ability to call programs, and
those may fit well in "common.py". Thus, "host.py" should not have a
dependency on "common.py" (but the other way around).

Optimally, we also move host.Host.ipa() to "common.py", because this
seems a more general helper method, and not something that belongs to
host.Host. "host.py" should probably not contain helpers that involve
invoking a program. For that, "common.py" would be the better place.
That is not done, but it shows that the better order of dependency is
that "common" imports "host", and not the other way around.