plugin: rename plugin implementations
Give the plugin modules a "plugin" prefix. The benfit is, that we
can easier see what plugins we have.

Also, note that the plugins are almost never directly imported, used or
referred to. The only place that imports these modules is during plugin
registration. Hence, the longer name is not a problem, because they are
only imported at one place.