plugin: add base class for plugin-specific "Task"
We have a "Task" base class, which is used for various purposes.

Tasks returned by a "pluginbase.Plugin", should be of a subtype
"PluginTask". The benefit is that from this task, we can get the plugin
back and in general there is stronger typing where plugins only deal
with plugin tasks.