trafficFlowTests: rename "_tc" attribute to "tc"
The "_tc" field is also accessed by "main.py". It is thus
not really private. Also, "TestConfig" is an immutable data
structure (or it should be immutable). Having no mutable
state, makes it less problematic to expose that state.

Rename "_tc" field to "tc", to not give the impression that
this is private when it is not.