testConfig: refactor TestConfig to make it better testable
Add a new argument "full_config" to TestConfig(). The purpose is for
testing, to inject a full config instead of having to write it to file
first.

Also, TestConfig would already in the constructor create the K8sClient
instances, and they would always use the kubeconfig files from "/root".
That makes it cumbersome to create a TestConfig instance for unit tests.

Instead,

- add a "mode_args" argument which allows the caller (the tests!) to
  overwrite the detection of the mode and kubeconfig.

- keep track of the kubeconfig strings, and create the K8sClient
  instances only when accessed, during TestConfig.client(). Note that
  this lazy construction of the K8sClient still leaves the object
  effectively immutable, because the K8sClient shall only be accessed via
  TestConfig.client(). So for the most part, you cannot notice that the
  K8sClient will only be created on first use.