testConfig: move testConfig.ClusterMode to tftbase
Enums are very trivial types (which is a good thing!). They have no state
and do nothing.

As such, adding them to "tftbase.py" is fine. "tftbase.py" is a fundamental module
that is used by most other modules. Usually, we would want that complex objects
have a limited scope and are defined in special modules, that are only used by
modules that need them. For a trivial enum, such hiding is unnecessary. If you
want to access an enum value, it should be easily possible from everywhere without
downsides.

Move the enum to "tftbase.py".