testConfig: parse and validate the config YAML
Add typed classes testConfig.ConfConfig tnd testConfig.Conf* to express
the content of the YAML.

When creating a TestConfig, the YAML gets parsed and we can access the
ConfConfig object at "tc.config".

These classes are immutable, as is the entire TestConfig object.

Also, perform strict validation/normalization and raise a useful error
message on configuration errors. The error message also contains a
yamlpath to where it is.

The benefit is that this validation/normalization happens already very
early, right when loading the configuration. An invalid configuration is
thus rejected earlier.

Another benefit is that all the interpretation/normalization is at one
place. For example, in YAML the field "sriov" maps to "pod_type" or
we accept "test_cases" as string or a list of enums.

The new configuration is not yet used. That comes next.