common: accept list argument for enum_convert_list()
The functions common.enum_convert() and common.enum_convert_list() exist
to accept a wider range of inputs (strings, numbers), and try to convert
that to the requested enum type. This exists as convenience to improve
usability.

Consider, that the value might come from YAML. In that case, for a list
of enums (that we would parse with enum_convert_list()) we accept
numbers and strings (and ranges in string form). For example,

  test_cases: "1,2,5-7"

Alternatively, also accept a list input, for example

  test_cases:
    - 1
    - 2
    - 5-7

Extend common.enum_convert_list() to also accept a list argument. When
a list if given, then each entry must either be a single enum value
(parsable via enum_convert()) or a string parsable as range (via
enum_convert_list()). Yes, the list item can itself contain string
ranges (see example), but nested Python lists are not supported.