testSettings: honor "persistent" setting from test config
This was previously ignored, and _server_test_to_persistent() would always
return false.

The comment "TODO: add logic to determine when this is required" makes it sound
as if this setting should be autodetected (based on the TestCaseType)?

The value from TestConfig always defaults to persistet=False. There is currently
no autodetection implemented. However, at least now it becomes configurable.

If a smarter default value should be chosen (e.g. detecting it), then this can be
added in the future.