common: add @strict_dataclass decorator
This is taken (and adjusted) from cluster-deployment-automation.

Normally, while dataclasses have type information about their fields,
Python does not check or enforce anything. Unless we want to get
surprised, we are well advised to only set types that we expect.

Add code for checking that the attributes of a dataclass have the
expected type. This is in form of a @strict_dataclass decorator, that
hocks __init__() and does additional checks after creation.