common: stricter error checking in dataclass_from_dict()
Better check for errors. But also handle for example Enum values, which
we must convert during dataclass_from_dict().

Previously, we would pass key values to the data class, which then had
to implement such conversion by implementing __init__() (or it silently
accepted a string for an enum).

Note that doing a best-effort/ignore-errors parsing would be more
complicated to implement correctly. Already previously, many errors were
silently ignored, but some were fatal (e.g. to initialize a field that
is itself a dataclass a dictionary had to be given). But best-effort
parsing is also not great. In the best case, it ignores invalid input
configuration. However it also will create partial/incomplete
configuration or even (as before) configuration that violates the type
annotation and expectation.

Now the parsing is strict, with similar checks as done with
@strict_dataclass.