tftbase: fix typing information for BaseOutput.result
This dictionary really can contain various types. It's not feasible to
list them all.

In fact, next we will tighten up dataclass_from_dict() to check the
types. The unit tests would fail then because they see other types in
the result dictionary.

While at it, use dict[] instead of Mapping[]. Mapping[] was used because
it's covariant for the value types, but it seems unnecessary for us.