pluginbase: fix cyclic import order of "testConfig" and "pluginbase"
We need a clear hierarchy and order in which modules should be imported.
Soon, we will want to import "pluginbase" in "testConfig", where it will
be useful for validating whether a plugin name is valid.

"pluginbase" is really quite a basic module. If it drags in the module
"perf", "task" and "testConfig", it cannot be included in those modules.

Instead:

- move "pluginbase.PluginTask" to "task.PluginTask". Note that
  "pluginbase" refers to "task.PluginTask" only for type annotations
  purposes. We can resolve that dependency, by forward declare the
  annotation as string (and import the module "if typing.TYPE_CHECKING").

- we refer to perf.PerfClient, perf.PertServer, testConfig.TestConfig
  from pluginbase module. But we only do so for type annotations of
  function arguments/return values. We can resolve that cycle by using
  strings instead of the actual types in the type annotation and import
  the module only with "if typing.TYPE_CHECKING".