netdev: add helper methods for getting links/addresses/routes
Taken and adjusted from cluster-deployment-automation. I think this is
the way, where our getters give typed (and validated) objects.

Also, by default any unexpected errors are treated as if no such entry
was present. Most callers would call these methods to check
whether/which links/addresses/routes are configured, and an error in
many cases can just be treated the same as if those entries don't exist.
If that is not desirable, then any failure can be turned into an
exception with the strict_parsing=True argument.

Kernel does not enforce that interface "names" (and many "names" in general) are
valid ASCII/UTF-8. We deal with that by using "errors='surrogateescape'".
So we get strings, but note that certain things may behave oddly. For
example, you cannot ifname.encode(errors="strict") such an interface
name, if it contains surrogate escapes. But you can print them,
convert back to binary (ifname.encode(errors="surrogateescape")), use in
path names and for the most part treat them as strings.