trafficFlowTest: fix podman-rm command line
`podman rm --time n` fails unless we also specify "--force".
If we do that, we don't need to call a separate "stop" first.

Also, currently Host.run() interprets the cmd string not as shell
script, so specifying two commands separated by semicolon is not going
to work. That probably should change, because interpreting the cmd as
shell script is a useful feature. Anyway, for now, this doesn't work.