task: extract initialize() method
Don't have the constructor write files or have other side effects.
Add an initialize() method.

The actual point here is to get a grasp of what these task
implementations are doing. Their __init__() methods are large, and set
various state in the object. Most of this state is unnecessary, and we
can do better.

The first step is to make __init__() simpler, and only set some values.

Also, this fixes an actual bug. It is Task.__init__() that initializes
self.template_args={}. That means, subclasses cannot add template
arguments before calling super()__init__(). On the other hand,
Perf.__init__() already renders the templates. That means, the
subclasses of Perf cannot set template_args, before the file gets
rendered. The proper fix for this is that __init__() only initializes
variables, and does not already render files.

Fixes: 0079e6bc3bef ('perf: fix IperfServer()/NetPerfServer() constructors with "persistent"')