perf: fix IperfServer()/NetPerfServer() constructors with "persistent"
We must first chain up the parent constructor before using
self.template_args. This was broken, if "persistent" were enabled.

Note that currently the parent constructor already writes the
templating. So while we avoid the crash, the templating still doesn't
work. That will be fixed later, by writing the templates due to an
explicit initialize() call.