testConfig: add  ConfigDescriptor to describe test context
Consider the problem.

"Task" would need to know the namespace. The namespace is part of the
configuration YAML, as

  tft:
    - name: "Test 1"
      namespace: "default"

But the entire configuration describes many tests and parts of tests (we
have the "tft" list, the "test_cases" list, the "connections" list).
Inside the Task, we already have access to the full TestConfig, but we
don't know which part of the entire configuration is relevant for the
current task (e.g.  the "tft" index, the "test_cases" index, etc).

We currently solve that by piggybacking each parameter from the top
level down. The "namespace" is at a high level, and it would have to be
passed through the layers down to the task.

However, as TestConfig is immutable and global we can pass this around,
if we only knew what part of the TestConfig we currently care about.

Add "ConfigDescriptor" for that. This is an immutable descriptor that
tells which subset of the tests is currently relevant.

If we pass a ConfigDescriptor instance to the Task, it can access the
entire TestConfig, but it also knows it's own context. For example, it
can get the correct namespace from cfg_descr.get_tft().namespace.

ConfigDescriptor basically bundles a few parameters that tell us where inside
the TestConfig we currently are. Previously, those had to be passed as
separate variables. Now it's all one immutable handle.

For example, see _run_test_tasks() which now can access the duration. We
no longer need to pass that as a separate parameter down the call stack.