task: pass TestSettings to Task for full configuration and context
TestSettings is an immutable object that contains information about the
current test sub run.

We have TestConfig that contains the entire configuration (of many tests and
cases). We have ConfigDescritor which gives additional information about
which part of the TestConfig is currently relevant.

We also have TestSettings, which contains a SelectTest (and indirectly
a TestConfig), and it contains additional context about the test.

All of these (TestSettings, TestConfig, ConfigDescriptor) are currently
immutable. Maybe in the future, it might make sense to allow modifying
TestSettings. But in any case, it is useful and fine that a task has
information about the TestSettings.

Pass it on.

Also, no need to keep the current TestSettings in the state of the
TrafficFlowTests object. It's specific to the current "sub" test that we
run, we can pass it on where we need it.