trafficFlowTests: don't keep the TestConfig as variable in TrafficFlowTests
Instead, pass it to the places where we need it.

It's also not actually useful, because a TestConfig can contain any
number of tests (the "tft" list). We still need to know, which test we
currently talk about. For that we have ConfigDescriptor, which already
contains the TestConfig.

No need to keep this data around on the TrafficFlowTest object.