trafficFlowTests: don't keep tft_output in TrafficFlowTests
Just pass it around, to where it's needed.

Not surprisingly, it turns out that this array is used only at a few
places. It's easier to review when the output is not parked in the
TrafficFlowTest instance, where it can be accessed by anybody (which
makes it harder to reason about where the state is used).