testSettings: make TestSettings a frozen dataclass
TestSettings is already immutable, in the sense that no code currently modifies
a TestSettings instance after creation.

Make it a frozen dataclass, which helps enforcing the immutablility via typing
hints and runtime checks.

Also, make the computed attributes to be property getters.  The benefit
is, that it is quite obvious that these getters don't modify the object.
More importantly, with the frozen data classes it would be cumbersome to
set these values as attributes.