pluginMeasurePower: return a json dictionary from TaskMeasurePower
Previously, the output of the TaskMeasurePower task is just a floating
point number.

Note that later on, generate_output() parses this as JSON. That still
works for a floating point value, however, that only gives a number.
Then we pass that result to PluginOutput(result=), which is typed to
be a dict[str,Any].

All the typing here is loose and it's quite hard to reason about what
happens. We probably should perform stricter parsing and not pass JSON
strings around. Regardless, follow at least some sort of typing, and get
a strdict instead of a floating point.