threading: add event_client_finished Event to TestSettings instead of SyncManager
Like event_server_alive, move the Event out of SyncManager and to the
TestSettings.

SyncManager tries to abstract this API, but it really just adds a "set"
method around event.set(). This abstraction makes it harder to grep for
all uses of "event_client_finished" and confuses more than it helps.

Also, the Event is local to the test. Attaching it to a global
SyncManager (and run "reset()" at the start of each test) is wrong.
Attach it to the TestSettings, which is test-specific data.