threading: add event_server_alive Event to TestSettings instead of SyncManager
SynaManager provides two methods (set_server_alive() and
wait_on_server_alive()). This mixes up unrelated synchronization
targets, while not providing any benefit.

Move the handling of server_alive out of SyncManager and add an Event
to TestSettings.

TestSettings are test-specific context, and the right place for the
synchronization parameter of the one test.

Now, you can grep for "event_server_alive" and find the few places that
are relevant for this topic.