evaluator: use evalConfig for parsing "eval-config.yaml" and accept no config
Use evalConfig module for parsing the configuration.

Also, note how "eval-config.yaml" only contains settings for two of the
existing TestType. If you were to use that file to evaluate a test for
another TestType, it would previously fail. Note how you can use
"generate_new_eval_config.py" to generate a YAML, however, that requires
that you have a .RESULT YAML to begin with (which were generated by
evaluator).

Instead, if the configuration does not contain a test, assume that the
Bitrate is NA and that the threshold is unset (pass all).