common: convert "int" values to "float" for dataclass_from_dict()
Note that mypy allows assigning an int at places where a float is
expected. So setting a dataclass field for type float to an int is fine.

@strict_dataclass and dataclass_check() chooses to reject that. I think
that makes sense, because although python accepts an int at most places
where a float is expected, there is still a difference. We probably
should catch that.

However, adjust dataclass_from_dict() to convert such an int to a float.
Also so that the result still validates according to our stricter
dataclass_check().