evaluator: add evalConfig module for representing (and parsing) the evaluator config YAML
I think it's good practice, that we use dataclasses for passing data
around. For example, the output of the test is a dataclass. Such
dataclasses we can serialize/deserialize conveniently (with error
checking) using common.dataclass_{to,from}_dict().

However, for user configuration, it is useful to put more effort into
this. The benefit is, that we can provide better error messages when
something is wrong and we can handle various ways for expressing the
same configuration (for example, a list of enums can be a single range
or a list of ranges with common.enum_convert_list()).

We already have a dedicated parser in testConfig.TestConfig. Add also
one for evalConfig.Config for the argument for evaluator.py.

This will be used later.