pluginValidateOffload: extract ethtool_stat_get_packets() and signal parse error
No longer emit exceptions when parsing invalid/unexpected ethtool data.
Instead, recognize the lack of the expected data, and forward the error
by setting "success" of the result to false.

An exception at this point is not a useful reaction. And we probably
should treat unexpected ethtool output not as a bug, but an error that
can happen at runtime (and needs to be handled). Handle the error by
setting the success to False, instead of emitting an exception.

Also, extract the code that parses "ethtool -S" output and move it
separate functions. Also add a unit tests.

Note that case2 (fallback to X_queue_..._packets) is not unit tested,
because I don't have such an output at hand. Also, the original code
searches for "_xdp_packets" suffix, but it's not clear to me where this
comes from (searching the kernel sources does not indicate that drivers
return such a field). In any case, when you encounter this case in the
wild, extend the unit test.