tftbase: move Bitrate class
This will be used later, when the test run already generates the Bitrate
result. It will then not only be used by Evaluator. Move the type.

Also, the tx/rx fields can be set to None, which means that the rate for
this part is not known. This will be useful for tests that don't capture
the bitrate in both directions or don't capture it at all.

Also, make it a frozen class and perform strict validation. For
convenience, you can pass `int` as tx/rx value, but the constructor will
convert them to floats.