task: let base implementation aggregate_output() handle flow-test result
Instead of having TestTypeHandler implementations override
_aggregate_output(), move the common code to Task.aggregate_output().

If a TestType implementation really wants to overwrite
Task.aggregate_output(), it still can. However, that should not be
necessary, because Task.aggregate_output() already all that should be
necessary. In particular, it honors IperfOutput's "success" field, which
allows Task.aggregate_output()  to do the right thing, without having a
per-test-type override.

Task._aggregate_output() still exists and some implementations use it
for logging. That part is still task specific, because the base class
does not know how to interpret the details of the task.