evaluator: enable plugin evaluation for all plugins
Of the 3 existing plugins, only "validate_offload" implemented
Plugin.eval_log(), which is called by Evaluator for collecting the
result. All our plugins should be collected in the results.

Instead of implementing eval_log() for all plugins, rework how it works.

The existing eval_log() implementation of PluginValidateOffload only
performed some static check on the PluginOutput. Now, already when
running the test, this check is run and PluginOutput.success is set.

Later, the default implementation of Plugin.eval_plugin_output()
(previously Plugin.eval_log()) considers PluginOutput.success. This way,
we have it implemented for all plugin types, they just need to set the
right PluginOutput.success.

Note that Plugin.eval_plugin_output() no longer does anything
plugin-specific. It only considers PluginOutput.success. In the future,
we may want that the valuation needs to consider the plugin, so we keep
this as a virtual function of Plugin.