validateOffload: always return a PluginOutput from task operation
We can return a BaseOutput. However, that is not propagated to the result output.
That means, it will be missing in the evaluatior.

For example, if you run TestCaseType.POD_TO_HOST_SAME_NODE, then
we would have returned a BaseOutput(success=True, msg='Hostbacked pod'),
and the result is not counted in the overall output:

  2024-07-02 08:11:57 INFO    [th:139829285459776]:   FlowTest results: Passed 5/5
  2024-07-02 08:11:57 INFO    [th:139829285459776]:   Plugin results: Passed 20/28

That is confusing, we would expect to see all plugins counted (and pass,
if they don't do anything useful).