simple: add a test type with a simple python tcp echo server/client
The purpose is to have an alternative test tool. In this case, it's a
relatively simple Python script where the client connects via TCP, sends
a message, and the server sends it back.

The point is also, that this script is relatively easily extendable,
either for manual testing or to add a new mode. Hence you can specify
additional command line arguments in the test configuration.

https://issues.redhat.com/browse/NHE-743