all: make the TFT_TOOLS_IMG configurable via environment variables
When developing a new feature, it can be cumbersome to test it, because
the test image only gets published after the feature is merged to
`main`.

Allow overriding the test image (which defaults to
"quay.io/wizhao/tft-tools:latest") by setting "TFT_TEST_IMAGE"
environment variable to a non-empty value.