all: allow overwriding the image pull police via environment variable
When making changes to the image, it can be important to force pulling the
latest image (TFT_TEST_IMAGE) from the repository.

Honor TFT_IMAGE_PULL_POLICY for that (for example, set it to "Always").
If unset, then we default to "IfNotPresent" unless "TFT_TEST_IMAGE"
environment variable is set.