testConfig: make kubeconfig/kubeconfig_infra configurable via TestConfig YAML
We detect the ClusterMode and the kubeconfigs based on whether we find
certain files in /root. That works, in common cases.

Allow to configure this explicitly, by setting the `kubeconfig` and
(optionally) `kubeconfig_infra` fields in the config YAML. If
`kubeconfig_infra` is set we determine the cluster mode as DPU.

https://github.com/wizhaoredhat/ocp-traffic-flow-tests/issues/25