common: allow None values in TestConfig
I didn't realize that YAML can easily have None/null values, e.g. with

   dict_key:

In almost all cases, we want to treat None as if the key is unset. In
that case, the parser in TestConfig gets it right already, because it
does

  v = vdata.pop("KEY", None)

and then proceeds to treat None as unset. What we just want.

I guess, there could be obscure cases, were we want to great None
differently from a missing key (but better not do that).

In any case, there is no reason to reject that outright. Either we can
accept an explicitly set value None as unset, or we could handle None.
In both cases, there is no problem with the user setting that.