manifests: fix rendering the "args" in Jinja2 templates
The arguments are a list. We should not set the replacement to
"args":"" for the empty list. Instead, set it to "args":[].

Also, fix "manifests/tools-pod.yaml.j2" to properly take a list argument
at that place.

So far this didn't matter, because "command":"/sbin/init" ignores all
arguments, and with `oc exec` the command and arguments are unused.
It will matter later.