pluginMeasurePower: improve error reporting of "measure_power" plugin
There is no need for individual logger.error() calls. We already call
self.run_oc_exec(), which will log an ERROR with the full output of the
command. That is more useful already.

Also, when running ipmitool or parsing its output fails, signal an error
and set a reason message.