testConfig: add ConfigDescriptor.get_{server,client}() helpers
The server/client is actually a list of entries, although currently
TestConfig enforces that this list always contains one entry.

If we would really support any number of entries, we would want to index
the server/client in the ConfigDescriptor instance, so we know which one we
talk about.

In a half-way step towards that, add get_server()/get_client accessors.
In the future, those shall become aware of the current index of the
server/client. For now, they just always return the single entry in the
list.