host: add Host.host_or_local() helper
This takes a Optional[Host] and either returns the Host or Host.local.
just a small convenience function for places where we have an Optional[Host],
where None means to fallback to Host.local. Which is common, especially
because today we have no remote implementation of Host, so all currently
existing callers always mean Host.local.