host: avoid duplicate decoding with "errors='surrogateescape'"
"errors='surrogateescape'" is a very useful way for handling binary
output of commands as if they were strings.

Adjust Host.run() to avoid trying to decode() such strings twice. We
first always decode with "errors='strict'", to see whether the input is
valid UTF-8. If it is, we don't need to decode a second time with
"errors='surrogateescape'". Add that optimization.