testType: make "command" template a list in Jinja2
Inside the pod custom resource, the "command" corresponds to the
ENTRYPOINT and the "args" to the CMD. Both are a list of strings, and
the invoked command is formed by concatenating both lists.

In practice, we currently always just provide one entry to "command"
(ENTRYPOINT) and expect the additional arguments in "args" (CMD).
So for our input, it was mostly fine that our templates had

  command:
    - "{{ command }}"

This seems an inconsistent mangling. The "command" is a list, so let the
code express the variable as list.

  command: {{ command }}