testType: don't overwrite the ENTRYPOINT with persistent mode
The Task base class sets the "command" Jinja2 variable to
"/usr/bin/container-entry-point.sh". This is a fine entry point, which
runs "tini" as a simple init process.

Don't let the subclasses overwrite "command" (that is, ENTRYPOINT). If
they do, they don't run with an init process. That works for some cases,
where the process handles SIGTERM and does not recursively spawn
subprocesses. However, in general, is is preferable to just use our
"tini" init process and the entry point that we commonly use.

So don't let subclasses overwrite "command", but only the args.